+++ modules/puss_nuklear/puss_module_nuklear_glfw3.c
@@ -311,7 +311,7 @@ nk_glfw3_mouse_button_callback(GLFWwindow* window, int button, int action, int m
         double dt = glfwGetTime() - glfw->last_button_click;
         if (dt > NK_GLFW_DOUBLE_CLICK_LO && dt < NK_GLFW_DOUBLE_CLICK_HI) {
             glfw->is_double_click_down = nk_true;
-            glfw->double_click_pos = nk_vec2(x, y);
+            glfw->double_click_pos = nk_vec2((float)x, (float)y);
         }
         glfw->last_button_click = glfwGetTime();
     } else glfw->is_double_click_down = nk_false;
@@ -453,7 +453,7 @@ nk_glfw3_new_frame(struct nk_glfw* glfw)
     nk_input_button(ctx, NK_BUTTON_LEFT, (int)x, (int)y, glfwGetMouseButton(win, GLFW_MOUSE_BUTTON_LEFT) == GLFW_PRESS);
     nk_input_button(ctx, NK_BUTTON_MIDDLE, (int)x, (int)y, glfwGetMouseButton(win, GLFW_MOUSE_BUTTON_MIDDLE) == GLFW_PRESS);
     nk_input_button(ctx, NK_BUTTON_RIGHT, (int)x, (int)y, glfwGetMouseButton(win, GLFW_MOUSE_BUTTON_RIGHT) == GLFW_PRESS);
-    nk_input_button(ctx, NK_BUTTON_DOUBLE, glfw->double_click_pos.x, glfw->double_click_pos.y, glfw->is_double_click_down);
+    nk_input_button(ctx, NK_BUTTON_DOUBLE, (int)(glfw->double_click_pos.x), (int)(glfw->double_click_pos.y), glfw->is_double_click_down);
     nk_input_scroll(ctx, glfw->scroll);
     nk_input_end(&glfw->ctx);
     glfw->text_len = 0;
@@ -535,8 +535,8 @@ static int glad_inited = 0;
 
 static int nk_glfw_window_create_lua(lua_State* L) {
 	const char* title = luaL_optstring(L, 1, "nuklear glfw window");
-	int width = luaL_optinteger(L, 2, 1024);
-	int height = luaL_optinteger(L, 3, 768);
+	int width = (int)luaL_optinteger(L, 2, 1024);
+	int height = (int)luaL_optinteger(L, 3, 768);
     GLFWwindow *win;
     struct nk_font_atlas *atlas;
 	struct nk_glfw* glfw = (struct nk_glfw*)lua_newuserdata(L, sizeof(struct nk_glfw));
